+++ physicsEngineDemo/src/RenderParticles.cpp
@@ -19,12 +19,12 @@ void RenderableParticle::display()
 
 	glBegin(GL_QUADS);
 	// Set the color for the particle
-	glColor3f(color[0], color[1], color[2]);
+	glColor3f((float)color[0], (float)color[1], (float)color[2]);
 	// Now render the particle
-	glVertex3f(position[0] - size, position[1] - size, position[2]);
-	glVertex3f(position[0] + size, position[1] - size, position[2]);
-	glVertex3f(position[0] + size, position[1] + size, position[2]);
-	glVertex3f(position[0] - size, position[1] + size, position[2]);
+	glVertex3f((float)(position[0] - size), (float)(position[1] - size), (float)position[2]);
+	glVertex3f((float)(position[0] + size), (float)(position[1] - size), (float)position[2]);
+	glVertex3f((float)(position[0] + size), (float)(position[1] + size), (float)position[2]);
+	glVertex3f((float)(position[0] - size), (float)(position[1] + size), (float)position[2]);
 	glEnd();
 
 	glColor4f(1, 1, 1, 1);
@@ -51,10 +51,10 @@ void CircleParticle::display()
 
 	PhysicsEngine::Vector3 position = physicsParticle->getPosition();
 
-	glColor3f(color[0], color[1], color[2]);
+	glColor3f((float)color[0], (float)color[1], (float)color[2]);
 	glPushMatrix();
-	glTranslatef(position[0], position[1], position[2]);
-	glutSolidSphere(size, 20.0, 20.0);
+	glTranslatef((float)position[0], (float)position[1], (float)position[2]);
+	glutSolidSphere((float)size, 20, 20);
 	glPopMatrix();
 }
